requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
Django==1.10
django-easy-pdf==0.1.1
django-multiforloop==0.2.1
gunicorn==19.8.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
lxml==4.2.1
Pillow==5.1.0
psycopg2==2.7.4
PyPDF2==1.26.0
reportlab==3.4.0
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
xhtml2pdf==0.2.2
